+++ lib/flutter_overlay_window.dart
@@ -3,3 +3,4 @@ library flutter_overlay_window;
 export 'package:flutter_overlay_window/src/overlay_window.dart';
 export 'package:flutter_overlay_window/src/overlay_config.dart';
 export 'package:flutter_overlay_window/src/models/overlay_position.dart';
+export 'package:flutter_overlay_window/src/models/overlay_orientation.dart';